[PAR-4830] patch checks for config / installer as central flow
