+++ Model/Config/Backend/EnableProductProtection.php
@@ -7,13 +7,11 @@
 namespace Extend\Integration\Model\Config\Backend;
 
 use Magento\Framework\App\Config\Value;
-use Extend\Integration\Setup\Model\AttributeSetInstaller;
 use Extend\Integration\Setup\Model\ProductInstaller;
 use Magento\Framework\App\Config\Storage\WriterInterface;
 
 class EnableProductProtection extends Value
 {
-    private AttributeSetInstaller $attributeSetInstaller;
     private ProductInstaller $productInstaller;
     private WriterInterface $writer;
 
@@ -22,7 +20,6 @@ class EnableProductProtection extends Value
         \Magento\Framework\Registry $registry,
         \Magento\Framework\App\Config\ScopeConfigInterface $config,
         \Magento\Framework\App\Cache\TypeListInterface $cacheTypeList,
-        AttributeSetInstaller $attributeSetInstaller,
         ProductInstaller $productInstaller,
         WriterInterface $writer,
         \Magento\Framework\Model\ResourceModel\AbstractResource $resource = null,
@@ -36,7 +33,6 @@ class EnableProductProtection extends Value
             $resource,
             $resourceCollection
         );
-        $this->attributeSetInstaller = $attributeSetInstaller;
         $this->productInstaller = $productInstaller;
         $this->writer = $writer;
     }
@@ -51,8 +47,7 @@ class EnableProductProtection extends Value
     {
         $isPPV2Enabled = (int) $this->getValue();
         if ($isPPV2Enabled === 1) {
-            $attributeSet = $this->attributeSetInstaller->createAttributeSet();
-            $this->productInstaller->createProduct($attributeSet);
+            $this->productInstaller->createProduct();
         } else {
             $this->productInstaller->deleteProduct();
             $this->writer->save(
